refactor: clean up api routes
