+++ lib/interfaces/spotify.js
@@ -1,4 +1,4 @@
-import { cookies } from "../services/cookies";
+import { cookies } from "../../services/cookies";
 
 /**
  * Class representing Spotify client
