+++ pages/api/v1/id.js
@@ -1,19 +1,15 @@
 import { NextApiRequest, NextApiResponse } from "next";
+import { getRequestInitOptions } from "../../../lib/helpers/index";
 
 /**
- * Gets request init options
+ * Parses data and gets user id
  *
- * @param {string} accessToken
+ * @param {Object} data
+ * @returns {string} id
  */
-function _getRequestInitOptions(accessToken) {
-  return {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-  };
+function _getId(data) {
+  const { id } = data;
+  return id;
 }
 
 /**
@@ -26,7 +22,7 @@ async function _getUser(accessToken, refresh) {
   const baseUrl = "https://api.spotify.com";
   const endpoint = "/v1/me";
   const api = baseUrl + endpoint;
-  const response = await fetch(api, _getRequestInitOptions(accessToken));
+  const response = await fetch(api, getRequestInitOptions(accessToken, "GET"));
   if (response.status === 401) {
     const baseUrl = "https://accounts.spotify.com";
     const endpoint = "/api/token";
@@ -46,17 +42,6 @@ async function _getUser(accessToken, refresh) {
     return data;
   }
   return await response.json();
-}
-
-/**
- * Parses data and gets user id
- *
- * @param {Object} data
- * @returns {string} id
- */
-function _getId(data) {
-  const { id } = data;
-  return id;
 }
 
 /**
