+++ pages/api/v1/recommendations.js
@@ -1,20 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-
-/**
- * Gets request init options
- *
- * @param {string} accessToken
- */
-function _getRequestInitOptions(accessToken) {
-  return {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-  };
-}
+import { getRequestInitOptions } from "../../../lib/helpers/index";
 
 /**
  * Gets recommmendations
@@ -24,7 +9,7 @@ async function _getRecommendations(accessToken, refresh, params) {
   const endpoint = "/v1/recommendations";
   const query = new URLSearchParams(params).toString();
   const api = baseUrl + endpoint + "?" + query;
-  const response = await fetch(api, _getRequestInitOptions(accessToken));
+  const response = await fetch(api, getRequestInitOptions(accessToken, "GET"));
   if (response.status === 401) {
     const baseUrl = "https://accounts.spotify.com";
     const endpoint = "/api/token";
