+++ pages/api/v1/search.js
@@ -1,4 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
+import { getRequestInitOptions } from "../../../lib/helpers/index";
 
 /**
  * Gets query params
@@ -14,19 +15,13 @@ function _getQueryParams(track) {
 }
 
 /**
- * Gets request init options
+ * Parses data and returns track uris
  *
- * @param {string} accessToken
+ * @param {Object} data
+ * @returns {string[]} uris
  */
-function _getRequestInitOptions(accessToken) {
-  return {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-  };
+function _getUris(data) {
+  return [data.tracks.items[0].uri];
 }
 
 /**
@@ -42,7 +37,7 @@ async function _search(accessToken, refresh, track) {
   const params = _getQueryParams(track);
   const query = new URLSearchParams(params).toString();
   const api = baseUrl + endpoint + "?" + query;
-  const response = await fetch(api, _getRequestInitOptions(accessToken));
+  const response = await fetch(api, getRequestInitOptions(accessToken, "GET"));
   if (response.status === 401) {
     const baseUrl = "https://accounts.spotify.com";
     const endpoint = "/api/token";
@@ -63,16 +58,6 @@ async function _search(accessToken, refresh, track) {
   }
   const json = await response.json();
   return json;
-}
-
-/**
- * Parses data and returns track uris
- *
- * @param {Object} data
- * @returns {string[]} uris
- */
-function _getUris(data) {
-  return [data.tracks.items[0].uri];
 }
 
 /**
